Convert Precursor::ActivationMethod to enum class
This converts the ActivationMethod enum in the Precursor class from a
plain C++ enum to a scoped enum class for improved type safety:

- Prevents implicit conversion to integers
- Requires explicit member qualification (e.g., Precursor::ActivationMethod::CID)
- Fixes Cython compilation issues with pyOpenMS bindings

Changes made:
- Modified Precursor.h: enum -> enum class, array sizes use static_cast
- Modified Precursor.cpp: Added static_assert for array size validation,
  updated array indexing to use static_cast<size_t>()
- Updated Precursor.pxd for Cython: cdef enum -> cdef enum class
- Updated all usages across codebase (format handlers, analysis modules,
  TOPP tools, GUI components, tests) to use qualified enum syntax
- Added compile-time assertions to ensure array sizes match enum count

+++ src/openms/source/FORMAT/HANDLERS/MzXMLHandler.cpp
@@ -1011,12 +1011,12 @@ namespace OpenMS::Internal
           }
           if (!precursor.getActivationMethods().empty())
           { // must not be empty, but technically only ETD, ECD, CID are allowed in mzXML 3.1
-            os << " activationMethod=\"" << Precursor::NamesOfActivationMethodShort[int(*(precursor.getActivationMethods().begin()))] << "\" ";
+            os << " activationMethod=\"" << Precursor::NamesOfActivationMethodShort[static_cast<size_t>(int(*(precursor.getActivationMethods().begin()))] << "\" ";
           } 
           else if (options_.getForceMQCompatability())
           { // a missing activation would make old MQ versions crash...
             OPENMS_LOG_WARN << "Warning: An MS2 scan does not have data on it's activation method. Using 'CID' as fallback!\n";
-            os << " activationMethod=\"" << Precursor::NamesOfActivationMethodShort[Precursor::ActivationMethod::CID] << "\" ";
+            os << " activationMethod=\"" << Precursor::NamesOfActivationMethodShort[static_cast<size_t>(Precursor::ActivationMethod::CID] << "\" ";
           }
 
           //m/z
